chore: domain event for kafka and jpaAuditingConfig  @louisphm091

+++ src/main/java/vn/com/routex/hub/booking/service/infrastructure/kafka/event/RouteOpenForBookingEvent.java
@@ -8,7 +8,7 @@ import java.time.OffsetDateTime;
 public record RouteOpenForBookingEvent(
         String routeId,
         String vehicleId,
-        Integer seatCount,
+        Long seatCount,
         String creator,
         OffsetDateTime assignedAt
 ) {
